fix: remove int32 cast

+++ mava/components/jax/executing/action_selection.py
@@ -85,7 +85,7 @@ class FeedforwardExecutorSelectAction(ExecutorSelectAction):
         executor.store.policies_info = {}
         for agent, observation in executor.store.observations.items():
             action_info, policy_info = executor.select_action(agent, observation)
-            executor.store.actions_info[agent] = np.int32(action_info)
+            executor.store.actions_info[agent] = action_info
             executor.store.policies_info[agent] = policy_info
 
     # Select action
